Adds colors to narrative display for transit legs
It changes background color of otp-Header leg and name of route.

It is shown only if transit legs have routeColor attribute. (It is assumed that
routeTextColor also exists)

It doesn't work nicely if routeColor is black. Because arrows and mode
icons are also black and are invisible if background is black.
Their color can't be changed with text color because they are images.

To fix this they would need to be icon font.

Fixes #40.