New mobile UI
When width is < 768 three tabs are shown:
Input, Map and plan output
Instead of sidebar and a map

It uses bootstrap tabs and some Javascript to move narrative from sidebar to own tab.
It probably needs some changes since it looks weird now. Sidebar is a little cut on both sides.

Fixes #41